disabled incorrect unit-test, fixed f32->i32

+++ inference-engine/tests/functional/inference_engine/transformations/convert_gather_0d_test.cpp
@@ -41,7 +41,7 @@ TEST(TransformationTests, ConvertGather0DStatic1) {
 
     {
         auto input = std::make_shared<opset1::Parameter>(element::f32, Shape{6, 12, 10, 24});
-        auto indices = std::make_shared<opset1::Parameter>(element::f32, Shape{15, 4, 20, 28});
+        auto indices = std::make_shared<opset1::Parameter>(element::i32, Shape{15, 4, 20, 28});
         auto axis_const = opset1::Constant::create(element::i64, Shape{}, {1});
         auto gather = std::make_shared<opset7::Gather>(input, indices, axis_const);
 
@@ -72,7 +72,7 @@ TEST(TransformationTests, ConvertGather0DStatic2) {
 
     {
         auto input = std::make_shared<opset1::Parameter>(element::f32, Shape{6, 12, 10, 24});
-        auto indices = std::make_shared<opset1::Parameter>(element::f32, Shape{});
+        auto indices = std::make_shared<opset1::Parameter>(element::i32, Shape{});
         auto axis_const = opset1::Constant::create(element::i64, Shape{}, {1});
         auto unsqueeze = std::make_shared<opset1::Unsqueeze>(indices, opset1::Constant::create(element::i64, Shape{1}, {0}));
         auto gather = std::make_shared<opset7::Gather>(input, unsqueeze, axis_const);
